# web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Understudy",
    "description": "Rehearse-then-dial voice recruiting on the Hunar Voice Agents API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/runs/{run_id}": {
      "get": {
        "tags": [
          "rehearsal"
        ],
        "summary": "A rehearsal run's status, scores, and per-case summaries",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/cases/{case_id}": {
      "get": {
        "tags": [
          "rehearsal"
        ],
        "summary": "One persona's full transcript, extraction, ground truth, metrics and failures",
        "operationId": "get_case_runs__run_id__cases__case_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/patch": {
      "post": {
        "tags": [
          "rehearsal"
        ],
        "summary": "Propose a prompt patch addressing this run's worst failures",
        "operationId": "propose_run_patch_runs__run_id__patch_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "persona_id": {
            "type": "string",
            "format": "uuid",
            "title": "Persona Id"
          },
          "archetype": {
            "type": "string",
            "title": "Archetype"
          },
          "transcript": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "extracted_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Result"
          },
          "ground_truth": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ground Truth"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "failures": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failures"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "persona_id",
          "archetype",
          "transcript",
          "extracted_result",
          "ground_truth",
          "metrics",
          "failures"
        ],
        "title": "CaseRead"
      },
      "CaseSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "persona_id": {
            "type": "string",
            "format": "uuid",
            "title": "Persona Id"
          },
          "archetype": {
            "type": "string",
            "title": "Archetype"
          },
          "turn_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Count"
          },
          "estimated_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "persona_id",
          "archetype",
          "turn_count",
          "estimated_seconds"
        ],
        "title": "CaseSummary",
        "description": "One row of RunRead.case_summaries \u2014 enough to pick a case to drill into, not the full\ntranscript (see GET /runs/{id}/cases/{case_id} for that)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "proposed_agent_prompt": {
            "type": "string",
            "title": "Proposed Agent Prompt"
          },
          "rationale": {
            "items": {},
            "type": "array",
            "title": "Rationale"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "resulting_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resulting Version Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "proposed_agent_prompt",
          "rationale",
          "accepted",
          "resulting_version_id"
        ],
        "title": "PatchRead"
      },
...
      "RunRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "agent_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Version Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scores": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scores"
          },
          "llm_calls": {
            "type": "integer",
            "title": "Llm Calls"
          },
          "cached_calls": {
            "type": "integer",
            "title": "Cached Calls"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "case_summaries": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "type": "array",
            "title": "Case Summaries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "agent_version_id",
          "status",
          "scores",
          "llm_calls",
          "cached_calls",
          "started_at",
          "finished_at",
          "error",
          "case_summaries"
        ],
        "title": "RunRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}